+++ packages/medical-domain/src/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, Diabeloop
+ * Copyright (c) 2022-2023, Diabeloop
  *
  * All rights reserved.
  *
@@ -29,8 +29,10 @@ import Datum from './domains/models/medical/datum.model'
 import Basal from './domains/models/medical/datum/basal.model'
 import BaseDatum from './domains/models/medical/datum/basics/base-datum.model'
 import Bolus from './domains/models/medical/datum/bolus.model'
+import { BolusSubtype } from './domains/models/medical/datum/enums/bolus-subtype.enum'
 import Cbg, { BgUnit, MGDL_UNITS, MMOLL_UNITS } from './domains/models/medical/datum/cbg.model'
 import ConfidentialMode from './domains/models/medical/datum/confidential-mode.model'
+import { DatumType } from './domains/models/medical/datum/enums/datum-type.enum'
 import DeviceParameterChange, { Parameter } from './domains/models/medical/datum/device-parameter-change.model'
 import { DurationValue } from './domains/models/medical/datum/basics/duration.model'
 import DurationUnit from './domains/models/medical/datum/enums/duration-unit.enum'
@@ -39,13 +41,19 @@ import Message from './domains/models/medical/datum/message.model'
 import PhysicalActivity from './domains/models/medical/datum/physical-activity.model'
 import Prescriptor from './domains/models/medical/datum/enums/prescriptor.enum'
 import PumpManufacturer from './domains/models/medical/datum/enums/pump-manufacturer.enum'
-import PumpSettings, { CgmConfig, DeviceConfig, ParameterConfig, PumpConfig } from './domains/models/medical/datum/pump-settings.model'
+import PumpSettings, {
+  CgmConfig,
+  DeviceConfig,
+  ParameterConfig,
+  PumpConfig
+} from './domains/models/medical/datum/pump-settings.model'
 import ReservoirChange from './domains/models/medical/datum/reservoir-change.model'
 import Smbg from './domains/models/medical/datum/smbg.model'
 import Source from './domains/models/medical/datum/enums/source.enum'
 import Unit from './domains/models/medical/datum/enums/unit.enum'
 import Upload from './domains/models/medical/datum/upload.model'
 import Wizard from './domains/models/medical/datum/wizard.model'
+import { WizardInputMealFat } from './domains/models/medical/datum/enums/wizard-input-meal-fat.enum'
 import ZenMode from './domains/models/medical/datum/zen-mode.model'
 import MedicalDataService from './domains/repositories/medical/medical-data.service'
 import { convertBG } from './domains/repositories/medical/datum/cbg.service'
@@ -53,8 +61,8 @@ import { BgClasses, defaultMedicalDataOptions, TimePrefs } from './domains/model
 import * as TimeService from './domains/repositories/time/time.service'
 
 export {
-  Basal, BaseDatum, BgClasses, BgUnit, Bolus, Cbg, CgmConfig, ConfidentialMode, Datum, DeviceConfig, DeviceParameterChange, DurationValue, DurationUnit, Meal, Message,
-  Parameter, ParameterConfig, PhysicalActivity, Prescriptor, PumpManufacturer, PumpConfig, PumpSettings, ReservoirChange, Smbg, Source, TimePrefs, Unit, Upload, Wizard, ZenMode,
+  Basal, BaseDatum, BgClasses, BgUnit, Bolus, BolusSubtype, Cbg, CgmConfig, ConfidentialMode, DatumType, Datum, DeviceConfig, DeviceParameterChange, DurationValue, DurationUnit, Meal, Message,
+  Parameter, ParameterConfig, PhysicalActivity, Prescriptor, PumpManufacturer, PumpConfig, PumpSettings, ReservoirChange, Smbg, Source, TimePrefs, Unit, Upload, Wizard, WizardInputMealFat, ZenMode,
   defaultMedicalDataOptions, TimeService, MGDL_UNITS, MMOLL_UNITS, convertBG
 }